refactor(lsp): split document symbol model
